# swagger.yaml
openapi: 3.0.0
info:
  title: Pimjo Blog API
  version: 1.0.0
  description: This simple blog API application allows users to create accounts, log in, and publish articles that are publicly visible. It provides robust authentication, thorough error handling, and complete testing, making it a reliable and user-friendly platform for blogging.
  contact:
    name: Pimjo Blog API Team
    url: https://example.com
    email: support@example.com
servers:
  - description: Development server
    url: http://localhost:8080
...
paths:
...
  /api/v1/articles:
    get:
      summary: Get all Articles
      description: Retrieve a list of articles.
      tags:
        - Articles
      parameters:
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/limitParam"
        - $ref: "#/components/parameters/sortTypeParam"
        - $ref: "#/components/parameters/sortByParam"
        - $ref: "#/components/parameters/searchParam"
      responses:
        "200":
          description: List of articles.
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                  data:
                    type: array
                    items:
                      allOf:
                        - type: object
                          properties:
                            author:
                              $ref: "#/components/schemas/AuthorDTO"
                        - $ref: "#/components/schemas/Article"

                  pagination:
                    $ref: "#/components/schemas/Pagination"
                  links:
                    type: object
                    properties:
                      self:
                        type: string
                        format: url
                        example: "/api/v1/articles?page=2&limit=10"
                      next:
                        type: string
                        format: url
                        example: "/api/v1/articles?page=3&limit=10"
                      prev:
                        type: string
                        format: url
                        example: "/api/v1/articles?page=1&limit=10"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/400"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/400"
    post:
      summary: Create a Article
      description: Create a new article.
      security:
        - bearerAuth: []
      tags:
        - Articles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - title
                - description
              properties:
                title:
                  type: string
                description:
                  type: string
      responses:
        "201":
          description: Article created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 201
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/Article"
                  links:
                    type: object
                    properties:
                      self:
                        type: string
                        format: url
                        example: "/api/v1/articles?page=2&limit=10"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/400"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/400"
  /api/v1/articles/{id}:
    get:
      summary: Get single Article
      description: Retrieve a article by ID.
      tags:
        - Articles
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the article to retrieve.
          schema:
            type: string
        - in: query
          name: expand
          description: expand comments array with
          schema:
            type: string
            example: "author comments"
      responses:
        "200":
          description: Article retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                  data:
                    type: object
                    allOf:
                      - type: object
                        properties:
                          author:
                            $ref: "#/components/schemas/AuthorDTO"
                          comments:
                            type: array
                            items:
                              $ref: "#/components/schemas/CommentDTO"
                      - $ref: "#/components/schemas/Article"

                  pagination:
                    $ref: "#/components/schemas/Pagination"
                  links:
                    type: object
                    properties:
                      self:
                        type: string
                        format: url
                        example: "/api/v1/articles?page=2&limit=10"
                      next:
                        type: string
                        format: url
                        example: "/api/v1/articles?page=3&limit=10"
                      prev:
                        type: string
                        format: url
                        example: "/api/v1/articles?page=1&limit=10"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/400"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/400"
    patch:
      summary: Update a Article
      description: Update a article by ID.
      security:
        - bearerAuth: []
      tags:
        - Articles
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the article to update.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        "200":
          description: Article updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/Article"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/400"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/400"
    delete:
      summary: Delete a Article
      description: Delete a article by ID.
      security:
        - bearerAuth: []
      tags:
        - Articles
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the article to delete.
          schema:
            type: string
      responses:
        "204":
          description: Article deleted successfully.
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/400"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/400"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    pageParam:
      in: query
      name: page
      description: Current page number (default 1).
      schema:
        type: integer
        format: int32
        example: 2
    limitParam:
      in: query
      name: limit
      description: Number of items per page (default 10).
      schema:
        type: integer
        format: int32
        example: 10
    sortTypeParam:
      in: query
      name: sort_type
      description: Type of sort, either asc or desc (default desc).
      schema:
        type: string
        enum: [asc, dsc]
        example: dsc
    sortByParam:
      in: query
      name: sort_by
      description: Property used for sorting, either updatedAt or title (default updatedAt).
      schema:
        type: string
        example: updatedAt
    searchParam:
      in: query
      name: search
      description: Search keyword for filtering items by name.
      schema:
        type: string
        example: title
...
  responses:
...
    "400":
      description: "Bad Request"
      content:
        "application/json":
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                format: int32
                example: 400
              error:
                type: string
                example: Bad Request
              message:
                type: string
                example: Invalid Parameters
              data:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                      example: email
                    message:
                      type: string
                      example: Invalid email
                    in:
                      type: string
                      enum: [query, body]
                      example: body
...
    "404":
      description: Resource Not Found
      content:
        "application/json":
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                format: int32
                example: 404
              error:
                type: string
                example: Not Found
              message:
                type: string
                example: Requested resource not found
  schemas:
    CommentDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 1
        description:
          type: string
          example: "Muhammad Sujon"
        author:
          $ref: "#/components/schemas/AuthorDTO"
      required: [id, name]
    AuthorDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 1
        name:
          type: string
          example: "Muhammad Sujon"
      required: [id, name]
    Timestamp:
      type: object
      properties:
        createdAt:
          description: Timestamp indicating when the item was created.
          type: string
          format: date-time
        updatedAt:
          description: Timestamp indicating when the item was last updated.
          type: string
          format: date-time
      required:
        - createdAt
        - updatedAt
    Pagination:
      type: object
      properties:
        page:
          type: integer
          format: int32
          example: 2
        limit:
          type: integer
          format: int32
          example: 10
        next:
          type: integer
          format: int32
          example: 3
        prev:
          type: integer
          format: int32
          example: 1
        totalPage:
          type: integer
          format: int32
          example: 5
        totalItems:
          type: integer
          format: int32
          example: 50
      required:
        - page
        - limit
        - totalPage
        - totalItems
...
    Article:
      allOf:
        - type: object
          required:
            - id
            - author
            - title
            - description
          properties:
            id:
              type: integer
              format: int32
              example: 1
            author:
              type: integer
              format: int32
              example: 1
            title:
              type: string
            description:
              type: string
              format: text
        - $ref: "#/components/schemas/Timestamp"
...